% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfoil_datta.R
\docType{data}
\name{airfoil_datta}
\alias{airfoil_datta}
\title{Scores of Group A and Group B}
\format{
A data frame with 60 rows and 2 variables:
\describe{
\item{group}{Participant's group, A or B.}
\item{score}{The participant's score in hypothetical task.}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Airfoil+Self-Noise}
}
\usage{
airfoil_datta
}
\description{
A data set with the scores of two groups.
}
\keyword{datasets}
