% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_sdr.R
\name{dr}
\alias{dr}
\title{Directional Regression}
\usage{
dr(x, y, nslices, d, ytype, lambda = 0)
}
\arguments{
\item{x}{a 'n x p' matrix of predictors; n sample size, p dimension}

\item{y}{a scalar response}

\item{nslices}{specify the number of slices to conduct;}

\item{d}{specify the reduced dimension}

\item{ytype}{specify the response as 'continuous' or 'categorical'}

\item{lambda}{a L2 or Tikonov regularizer for the sample covariance matrix;
default is '0', i.e. no regularization}
}
\value{
A list containing both the estimate and candidate matrix.
\itemize{
\item beta - A 'pxd' matrix that estimates a basis for the central subspace.
\item cand_mat - The candidate matrix for DR; this is used in other functions
for order determination.
}
}
\description{
This conducts directional regression (DR) as in Li (2018) with modifications
to improve speed and to allow for the option of standardizing and regularizing
}
\details{
Standardizing is the default as it is necessary for recovering the properly scaled
central subspace. However, in certain contexts, the standardization is not necessary,
and so we leave this option open to the practitioner.

The L2-regularization option corresponds to the SIR regularization idea
by Zhang et al.(2005). While they do not apply the idea to SAVE, we find that the
context is analogous and that such a regularization works.
}
