% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_plots.R
\name{ggplot_fsdr}
\alias{ggplot_fsdr}
\title{Produces ggplot objects for a one dimensional response and up to 2 
sufficient predictors.}
\usage{
ggplot_fsdr(
  y_datta,
  x_datta,
  y_on_axis = F,
  ytype = "multinomial",
  size = 1,
  h_lim = NULL,
  v_lim = NULL,
  h_lab = NULL,
  v_lab = NULL,
  main_lab = NULL,
  show_legend = T,
  y_colors = NULL,
  y_symbols = NULL
)
}
\description{
Produces ggplot objects for a one dimensional response and up to 2 
sufficient predictors.
}
