\name{linearsdr-package}
\alias{linearsdr-package}
\alias{linearsdr}
\docType{package}
\title{
\packageTitle{linearsdr}
}
\description{
\packageDescription{linearsdr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linearsdr}
\packageIndices{linearsdr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{linearsdr}

Maintainer: \packageMaintainer{linearsdr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
