% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pendigits_datta.R
\docType{data}
\name{pendigits_datta}
\alias{pendigits_datta}
\title{Scores of Group A and Group B}
\format{
A data frame with 60 rows and 2 variables:
\describe{
\item{group}{Participant's group, A or B.}
\item{score}{The participant's score in hypothetical task.}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Pen-Based+Recognition+of+Handwritten+Digits}
}
\usage{
pendigits_datta
}
\description{
A data set with the scores of two groups.
}
\keyword{datasets}
